Embed Z-Wave platforms (#20810)
