+++ homeassistant/components/zwave/binary_sensor.py
